pyproject.toml
[project]
name = "shefport"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.22.0",
    "alembic>=1.17.2",
    "apscheduler>=3.11.1",
    "bs4>=0.0.2",
    "loguru>=0.7.3",
    "mypy>=1.18.2",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "psycopg2>=2.9.11",
    "pymorphy3>=2.0.6",
    "python-dotenv>=1.1.1",
    "requests>=2.32.5",
    "ruff>=0.14.7",
    "selenium>=4.38.0",
    "setuptools>=80.9.0",
    "sqlalchemy>=2.0.44",
    "xlrd>=2.0.2",
    "yadisk>=3.4.0",
]
